requirements.txt
opencv-python
pandas
numpy
tqdm
dtaidistance
sklearn
multiprocessing
matplotlib-python